Adds darknet (for yolov4) in place of Plasticity.
This will allow for much higher accuracy and speed in object detection.
It also ends a personal era of trying to code everything from scratch as
Plasticity was my own neural network framework. In 1 day of Bazel
dependency wrangling, I managed to replace 3 years of work with someone
else's framework which is much better. That's a personal lesson for me.